Fix up some integer types in crypto/asn1
tag and utype are always accessed as int, so make the structs match.
Boolean ASN1_ITEMs put an ASN1_BOOLEAN in it->size, so add a cast. Also
fix the time set_string functions to call the underlying CBS parser
directly, so they don't need to put a strlen into an int.

Bug: 516
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/55449
Commit-Queue: Bob Beck <bbe@google.com>

+++ crypto/asn1/a_strex.c
@@ -121,7 +121,7 @@ static int do_esc_char(uint32_t c, unsigned long flags, char *do_quotes,
     return maybe_write(out, &u8, 1) ? 1 : -1;
   }
 
-  int len = strlen(buf);
+  int len = (int)strlen(buf);  // |buf| is guaranteed to be short.
   return maybe_write(out, buf, len) ? len : -1;
 }
