+++ include/openssl/asn1t.h
@@ -349,7 +349,7 @@ typedef struct ASN1_TLC_st ASN1_TLC;
 
 struct ASN1_TEMPLATE_st {
 uint32_t flags;		/* Various flags */
-long tag;			/* tag, not used if no tagging */
+int tag;			/* tag, not used if no tagging */
 unsigned long offset;		/* Offset of this field in structure */
 const char *field_name;		/* Field name */
 ASN1_ITEM_EXP *item;		/* Relevant ASN1_ITEM or ASN1_ADB */
@@ -455,7 +455,7 @@ struct ASN1_ADB_TABLE_st {
 
 struct ASN1_ITEM_st {
 char itype;			/* The item type, primitive, SEQUENCE, CHOICE or extern */
-long utype;			/* underlying type */
+int utype;			/* underlying type */
 const ASN1_TEMPLATE *templates;	/* If SEQUENCE or CHOICE this contains the contents */
 long tcount;			/* Number of templates if SEQUENCE or CHOICE */
 const void *funcs;		/* functions that handle this type */
